Fix PR comments from Wei
