+++ tests/api/ask_astro/rest/controllers/test_post_request.py
@@ -1,21 +1,6 @@
 from unittest.mock import AsyncMock, patch
 
 import pytest
-from sanic import Sanic
-from sanic_testing import TestManager
-
-sanitized_name = __name__.replace(".", "_")
-
-
-@pytest.fixture
-def app():
-    """Fixture to create a new Sanic application for testing the POST request handler."""
-    app_instance = Sanic(sanitized_name)
-    TestManager(app_instance)
-    from ask_astro.rest.controllers import register_routes
-
-    register_routes(app_instance)
-    return app_instance
 
 
 @pytest.mark.asyncio
